samples: modules: canopennode: set bootloader in sample.yaml
Set the bootloader choice for the
sample.modules.canopennode.program_download test case in the sample.yaml
file instead of sysbuild.conf. The reason for this change is to avoid
building MCUboot by default when using the --sysbuild option, as it may
not be compatible with all platforms. This is currently causing issues
on platforms like nRF54H20 where sysbuild is always used.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>