include: adc: Fix initialization of .differential in ADC_CHANNEL_CFG_DT
This is a follow-up to commit ff0f389d0b26999443444b0e37c4d66d4a845eff.

When the `zephyr,differential` property is used together with a driver
that selects `ADC_CONFIGURABLE_INPUTS`, the `differential` field is
initialized twice what causes a compilation error in C++. Fix this by
refactoring the logic around initialization of `.differential`.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>